Fix rendering issue on iphone5C
it's caused by the model view projection matrix's precision, which should be set to highp instead of mediump

+++ CubeEngine/CubeEngine/ShaderResources/BaseLightEffect.vert
@@ -1,6 +1,6 @@
 
 attribute lowp vec3 VertexNormal;
-uniform lowp mat3 NormalMatrix;
+uniform highp mat3 NormalMatrix;
 uniform lowp vec3 EyeDirection; // in eye space
 
 varying lowp vec3 Normal;
